+++ test/commentParserToESTree.js
@@ -1,52 +1,69 @@
 import {commentParserToESTree} from '../src/commentParserToESTree.js';
 import {parseComment} from '../src/parseComment.js';
 
+const singleLineWithTag = {
+  type: 'JsdocBlock',
+  delimiter: '/**',
+  description: '',
+  descriptionLines: [],
+  end: '*/',
+  endLine: 0,
+  lastDescriptionLine: 0,
+  lineEnd: '',
+  postDelimiter: ' ',
+  tags: [
+    {
+      delimiter: '',
+      description: '',
+      descriptionLines: [],
+      lineEnd: '',
+      name: '',
+      parsedType: {
+        type: 'JsdocTypeName',
+        value: 'string'
+      },
+      postDelimiter: '',
+      postName: '',
+      postTag: ' ',
+      postType: ' ',
+      tag: 'type',
+      type: 'JsdocTag',
+      rawType: 'string',
+      start: '',
+      typeLines: [
+        {
+          delimiter: '',
+          postDelimiter: '',
+          rawType: 'string',
+          start: '',
+          type: 'JsdocTypeLine'
+        }
+      ]
+    }
+  ]
+};
+
 describe('commentParserToESTree', function () {
   it('handles single line jsdoc comment with tag', () => {
     const parsedComment = parseComment({
       value: `* @type {string} `
     });
 
-    const ast = commentParserToESTree(parsedComment, 'javascript');
+    const ast = commentParserToESTree(parsedComment, 'jsdoc');
 
-    expect(ast).to.deep.equal({
-      type: 'JsdocBlock',
-      delimiter: '/**',
-      description: '',
-      descriptionLines: [],
-      end: '*/',
-      endLine: 0,
-      lastDescriptionLine: 0,
-      lineEnd: '',
-      postDelimiter: ' ',
-      tags: [
-        {
-          delimiter: '/**',
-          description: '',
-          descriptionLines: [],
-          lineEnd: '',
-          name: '',
-          parsedType: null,
-          postDelimiter: ' ',
-          postName: '',
-          postTag: ' ',
-          postType: ' ',
-          tag: 'type',
-          type: 'JsdocTag',
-          rawType: 'string',
-          start: '',
-          typeLines: [
-            {
-              delimiter: '/**',
-              postDelimiter: ' ',
-              rawType: 'string',
-              start: '',
-              type: 'JsdocTypeLine'
-            }
-          ]
-        }
-      ]
+    expect(ast).to.deep.equal(singleLineWithTag);
+  });
+
+  it('throws with bad injected type in single line jsdoc comment', () => {
+    const parsedComment = parseComment({
+      value: `* @type {string} `
     });
+
+    parsedComment.source[0].tokens.type = 'badValue<';
+
+    expect(() => {
+      commentParserToESTree(parsedComment, 'jsdoc');
+    }).to.throw("No parslet found for token: 'EOF'");
   });
 
   it('handles multi line jsdoc comment beginning on line 0', () => {
@@ -55,7 +72,7 @@ describe('commentParserToESTree', function () {
 `
     });
 
-    const ast = commentParserToESTree(parsedComment, 'javascript');
+    const ast = commentParserToESTree(parsedComment, 'jsdoc');
 
     expect(ast).to.deep.equal({
       type: 'JsdocBlock',
@@ -69,13 +86,16 @@ describe('commentParserToESTree', function () {
       postDelimiter: ' ',
       tags: [
         {
-          delimiter: '/**',
+          delimiter: '',
           description: '',
           descriptionLines: [],
           lineEnd: '',
           name: '',
-          parsedType: null,
-          postDelimiter: ' ',
+          parsedType: {
+            type: 'JsdocTypeName',
+            value: 'string'
+          },
+          postDelimiter: '',
           postName: '',
           postTag: ' ',
           postType: '',
@@ -85,8 +105,8 @@ describe('commentParserToESTree', function () {
           start: '',
           typeLines: [
             {
-              delimiter: '/**',
-              postDelimiter: ' ',
+              delimiter: '',
+              postDelimiter: '',
               rawType: 'string',
               start: '',
               type: 'JsdocTypeLine'
@@ -103,7 +123,7 @@ describe('commentParserToESTree', function () {
  * @type {string}`
     });
 
-    const ast = commentParserToESTree(parsedComment, 'javascript');
+    const ast = commentParserToESTree(parsedComment, 'jsdoc');
 
     expect(ast).to.deep.equal({
       type: 'JsdocBlock',
@@ -122,7 +142,10 @@ describe('commentParserToESTree', function () {
           descriptionLines: [],
           lineEnd: '',
           name: '',
-          parsedType: null,
+          parsedType: {
+            type: 'JsdocTypeName',
+            value: 'string'
+          },
           postDelimiter: ' ',
           postName: '',
           postTag: ' ',
@@ -133,10 +156,10 @@ describe('commentParserToESTree', function () {
           start: ' ',
           typeLines: [
             {
-              delimiter: '*',
-              postDelimiter: ' ',
+              delimiter: '',
+              postDelimiter: '',
               rawType: 'string',
-              start: ' ',
+              start: '',
               type: 'JsdocTypeLine'
             }
           ]
@@ -152,7 +175,7 @@ describe('commentParserToESTree', function () {
  *`
     });
 
-    const ast = commentParserToESTree(parsedComment, 'javascript');
+    const ast = commentParserToESTree(parsedComment, 'jsdoc');
 
     expect(ast).to.deep.equal({
       type: 'JsdocBlock',
@@ -171,7 +194,10 @@ describe('commentParserToESTree', function () {
           descriptionLines: [],
           lineEnd: '',
           name: '',
-          parsedType: null,
+          parsedType: {
+            type: 'JsdocTypeName',
+            value: 'string'
+          },
           postDelimiter: ' ',
           postName: '',
           postTag: ' ',
@@ -182,10 +208,10 @@ describe('commentParserToESTree', function () {
           start: ' ',
           typeLines: [
             {
-              delimiter: '*',
-              postDelimiter: ' ',
+              delimiter: '',
+              postDelimiter: '',
               rawType: 'string',
-              start: ' ',
+              start: '',
               type: 'JsdocTypeLine'
             }
           ]
@@ -195,17 +221,149 @@ describe('commentParserToESTree', function () {
   });
 
   it(
+    'handles multi line jsdoc comment with tag and multiline type',
+    () => {
+      const parsedComment = parseComment({
+        value: `*
+ * @type {{
+ *   a: string;
+     b: number;
+ *   c: null
+ * }} SomeName
+ *`
+      });
+
+      const ast = commentParserToESTree(parsedComment, 'jsdoc');
+
+      expect(ast).to.deep.equal({
+        delimiter: '/**',
+        description: '',
+        descriptionLines: [],
+        end: '*/',
+        endLine: 6,
+        lastDescriptionLine: 1,
+        lineEnd: '',
+        postDelimiter: '',
+        tags: [
+          {
+            delimiter: '*',
+            description: '',
+            descriptionLines: [],
+            lineEnd: '',
+            name: 'SomeName',
+            parsedType: {
+              elements: [
+                {
+                  key: 'a',
+                  meta: {
+                    hasLeftSideExpression: false,
+                    quote: undefined
+                  },
+                  optional: false,
+                  readonly: false,
+                  right: {
+                    type: 'JsdocTypeName',
+                    value: 'string'
+                  },
+                  type: 'JsdocTypeKeyValue'
+                },
+                {
+                  key: 'b',
+                  meta: {
+                    hasLeftSideExpression: false,
+                    quote: undefined
+                  },
+                  optional: false,
+                  readonly: false,
+                  right: {
+                    type: 'JsdocTypeName',
+                    value: 'number'
+                  },
+                  type: 'JsdocTypeKeyValue'
+                },
+                {
+                  key: 'c',
+                  meta: {
+                    hasLeftSideExpression: false,
+                    quote: undefined
+                  },
+                  optional: false,
+                  readonly: false,
+                  right: {
+                    type: 'JsdocTypeNull'
+                  },
+                  type: 'JsdocTypeKeyValue'
+                }
+              ],
+              meta: {
+                separator: 'semicolon'
+              },
+              type: 'JsdocTypeObject'
+            },
+            postDelimiter: ' ',
+            postName: '',
+            postTag: ' ',
+            postType: '',
+            rawType: '{\n  a: string;\nb: number;\n  c: null\n}',
+            start: ' ',
+            tag: 'type',
+            type: 'JsdocTag',
+            typeLines: [
+              {
+                delimiter: '',
+                postDelimiter: '',
+                rawType: '{',
+                start: '',
+                type: 'JsdocTypeLine'
+              },
+              {
+                delimiter: '*',
+                postDelimiter: ' ',
+                rawType: '  a: string;',
+                start: ' ',
+                type: 'JsdocTypeLine'
+              },
+              {
+                delimiter: '',
+                postDelimiter: '',
+                rawType: 'b: number;',
+                start: '     ',
+                type: 'JsdocTypeLine'
+              },
+              {
+                delimiter: '*',
+                postDelimiter: ' ',
+                rawType: '  c: null',
+                start: ' ',
+                type: 'JsdocTypeLine'
+              },
+              {
+                delimiter: '*',
+                postDelimiter: ' ',
+                rawType: '}',
+                start: ' ',
+                type: 'JsdocTypeLine'
+              }
+            ]
+          }
+        ],
+        type: 'JsdocBlock'
+      });
+    }
+  );
+
+  it(
     'handles multi line jsdoc comment with tag and multiline description',
     () => {
       const parsedComment = parseComment({
         value: `*
  * @param {string} Some
- * multi-line
+ * multiline
  description
  *`
       });
 
-      const ast = commentParserToESTree(parsedComment, 'javascript');
+      const ast = commentParserToESTree(parsedComment, 'jsdoc');
 
       expect(ast).to.deep.equal({
         type: 'JsdocBlock',
@@ -220,13 +378,13 @@ describe('commentParserToESTree', function () {
         tags: [
           {
             delimiter: '*',
-            description: 'multi-line\ndescription',
+            description: 'multiline\ndescription',
             descriptionLines: [
               {
-                delimiter: '*',
-                description: 'multi-line',
-                postDelimiter: ' ',
-                start: ' ',
+                delimiter: '',
+                description: 'multiline',
+                postDelimiter: '',
+                start: '',
                 type: 'JsdocDescriptionLine'
               },
               {
@@ -239,7 +397,10 @@ describe('commentParserToESTree', function () {
             ],
             lineEnd: '',
             name: 'Some',
-            parsedType: null,
+            parsedType: {
+              type: 'JsdocTypeName',
+              value: 'string'
+            },
             postDelimiter: ' ',
             postName: '',
             postTag: ' ',
@@ -250,10 +411,10 @@ describe('commentParserToESTree', function () {
             start: ' ',
             typeLines: [
               {
-                delimiter: '*',
-                postDelimiter: ' ',
+                delimiter: '',
+                postDelimiter: '',
                 rawType: 'string',
-                start: ' ',
+                start: '',
                 type: 'JsdocTypeLine'
               }
             ]
@@ -269,28 +430,28 @@ describe('commentParserToESTree', function () {
       const parsedComment = parseComment({
         value: `*
  * Some
- * multi-line
+ * multiline
  description
  *`
       });
 
-      const ast = commentParserToESTree(parsedComment, 'javascript');
+      const ast = commentParserToESTree(parsedComment, 'jsdoc');
 
       expect(ast).to.deep.equal({
         type: 'JsdocBlock',
         delimiter: '/**',
-        description: 'Some\nmulti-line\ndescription',
+        description: 'Some\nmultiline\ndescription',
         descriptionLines: [
           {
-            delimiter: '*',
+            delimiter: '',
             description: 'Some',
-            postDelimiter: ' ',
-            start: ' ',
+            postDelimiter: '',
+            start: '',
             type: 'JsdocDescriptionLine'
           },
           {
             delimiter: '*',
-            description: 'multi-line',
+            description: 'multiline',
             postDelimiter: ' ',
             start: ' ',
             type: 'JsdocDescriptionLine'
@@ -312,4 +473,45 @@ describe('commentParserToESTree', function () {
       });
     }
   );
+
+  it('handles multi line jsdoc comment with tag and name but no type', () => {
+    const parsedComment = parseComment({
+      value: `*
+ * @param TagNameNoType
+ *`
+    });
+
+    const ast = commentParserToESTree(parsedComment, 'jsdoc');
+
+    expect(ast).to.deep.equal({
+      delimiter: '/**',
+      description: '',
+      descriptionLines: [],
+      end: '*/',
+      endLine: 2,
+      lastDescriptionLine: 1,
+      lineEnd: '',
+      postDelimiter: '',
+      tags: [
+        {
+          delimiter: '*',
+          description: '',
+          descriptionLines: [],
+          lineEnd: '',
+          name: 'TagNameNoType',
+          parsedType: null,
+          postDelimiter: ' ',
+          postName: '',
+          postTag: ' ',
+          postType: '',
+          rawType: '',
+          start: ' ',
+          tag: 'param',
+          type: 'JsdocTag',
+          typeLines: []
+        }
+      ],
+      type: 'JsdocBlock'
+    });
+  });
 });